refactor benefit handler
